Move getDeadline() and getDescription() to TaskList
