+++ data/duke.txt
@@ -1,2 +1,3 @@
 1.T|Done a 
 2.T|Done a 
+3.E|NotDone a (sk: 10 Oct 2020)
