+++ src/main/java/Duke.java
@@ -25,77 +25,6 @@ public class Duke {
         storage = new Storage(filepath);
     }
 
-
-    /**
-     * A method to get the description from a task entered by the user
-     *
-     * @param arr The array that contains strings from the user input
-     * @return A string containing the description only
-     */
-    public static String getDescription(String[] arr) {
-        String str = "";
-        for(int i = 1; i < arr.length; i++) {
-            if (!(arr[i].charAt(0) == '/')) {
-                str += arr[i] + " ";
-            }
-            else {
-                break;
-            }
-        }
-        return str;
-    }
-
-    /**
-     * A method to get the deadline from a task entered by the user
-     *
-     * @param arr The array that contains strings from the user input
-     * @return A string containing the deadline only
-     */
-    public static String getDeadline(String[] arr) throws DateTimeParseException {
-        String str = "";
-        boolean bool = false;
-        for(int i = 1; i < arr.length; i++) {
-            if (!bool) {
-                if (arr[i].charAt(0) == '/') {
-                    bool = true;
-                    str += arr[i].substring(1) + ": ";
-                }
-            } else {
-                if (arr[i].length() == 10 && arr[i].charAt(4) == '-' && arr[i].charAt(7) == '-') {
-                    LocalDate ld = LocalDate.parse(arr[i]);
-                    str += ld.getDayOfMonth() + " "
-                            + ld.getMonth().getDisplayName(TextStyle.SHORT, Locale.ENGLISH) + " "
-                                + ld.getYear();
-                    if (i != arr.length - 1) {
-                        str += " ";
-                    }
-                } else if (arr[i].length() == 5 && arr[i].charAt(2) == ':') {
-                    LocalTime lt = LocalTime.parse(arr[i]);
-                    int hour = lt.getHour();
-                    String hourSuffix = hour < 12 ? "AM" : "PM";
-                    if (hour == 0) {
-                        hour = 12;
-                    } else if (hour > 12) {
-                        hour -= 12;
-                    }
-                    int minute = lt.getMinute();
-                    String minutePrefix = minute < 10 ? "0" : "";
-                    str += hour + ":" + minutePrefix + minute + " " + hourSuffix;
-                    if (i != arr.length - 1) {
-                        str += " ";
-                    }
-
-                } else {
-                    str += arr[i];
-                    if (i != arr.length - 1) {
-                        str += " ";
-                    }
-                }
-            }
-        }
-        return str;
-    }
-
     /**
      * A method to check if a string is an integer or not
      *
@@ -166,19 +95,19 @@ public class Duke {
                     if (arr.length < 2) {
                         throw new EmptyDescriptionException("Missing description / date");
                     }
-                    taskList.addTask(new Todo(getDescription(arr)));
+                    taskList.addTask(new Todo(TaskList.getDescription(arr)));
                     taskList.printAddTask();
                 } else if (arr[0].equals("deadline")) {
                     if (arr.length < 2) {
                         throw new EmptyDescriptionException("Missing description / date");
                     }
-                    taskList.addTask(new Deadline(getDescription(arr), getDeadline(arr)));
+                    taskList.addTask(new Deadline(TaskList.getDescription(arr), TaskList.getDeadline(arr)));
                     taskList.printAddTask();
                 } else if (arr[0].equals("event")) {
                     if (arr.length < 2) {
                         throw new EmptyDescriptionException("Missing description / date");
                     }
-                    taskList.addTask(new Event(getDescription(arr), getDeadline(arr)));
+                    taskList.addTask(new Event(TaskList.getDescription(arr), TaskList.getDeadline(arr)));
                     taskList.printAddTask();
                 } else if (input.equals("list")) {
                     taskList.displayList();
